Notes: implement pitch bends
This also renames the `start` and `end` fields of the Notes::Event struct
to respectively `startTime` and `endTime`, as well as adds two extra fields:
`startFrame` and `endFrame` to avoid back and forth conversions.

Contrary to the original basic pitch implementation the note events are now
always sorted regardless of parameters. Tests have been updated to reflect this.

Signed-off-by: Tibor Vass <teabee89@gmail.com>