Source audio manager (#38)
- New SourceAudioManager class that handles:
    - Recording of incoming audio at native and Basic Pitch sample rate.
    - Audio file drop
    - Audio thumbnail management.

The recording is now done directly to 2 wav files (native sample rate and basic pitch sample rate) on separate threads thanks to juce::ThreadedWriter. There is no 3 minutes limit anymore. The saved files are deleted on a click of clear button.

A timer is no longer used for the thumbnail, we simply use a changeListener.

- New AudioUtils namespace handling audio file loading and resampling with static functions.
- Fix bug in Resampler (there were some pops in 22.05 kHz audio due to bad streaming resampling).